Moved functions for working with macros and their differences to a separate file.
